requirements.txt
streamlit
requests
beautifulsoup4
mysql-connector-python
sentence-transformers
pymilvus
configparser
